+++ solutions/PE193.py
@@ -67,7 +67,7 @@ class Problem193:
         self.limit = self.n - 1
         self.total = self.n - 1
         self.ls_primes = primes((self.n ** 0.5) + 1)
-        self.ls_sq_primes = [p * p for p in self.ls_primes]
+        self.ls_sq_primes = [int(p * p) for p in self.ls_primes]
         self.num_primes = len(self.ls_primes)
         self.inclusion_exclusion_helper(odd_even=-1, prev_prod=1, prime_index=0, next_prod=4)
         return self.total
